+++ libcxx/test/std/utilities/utility/pairs/pairs.spec/make_pair.pass.cpp
@@ -21,7 +21,7 @@ int main()
 {
     {
         typedef std::pair<int, short> P1;
-        P1 p1 = std::make_pair(3, 4);
+        P1 p1 = std::make_pair(3, static_cast<short>(4));
         assert(p1.first == 3);
         assert(p1.second == 4);
     }
@@ -29,13 +29,13 @@ int main()
 #if TEST_STD_VER >= 11
     {
         typedef std::pair<std::unique_ptr<int>, short> P1;
-        P1 p1 = std::make_pair(std::unique_ptr<int>(new int(3)), 4);
+        P1 p1 = std::make_pair(std::unique_ptr<int>(new int(3)), static_cast<short>(4));
         assert(*p1.first == 3);
         assert(p1.second == 4);
     }
     {
         typedef std::pair<std::unique_ptr<int>, short> P1;
-        P1 p1 = std::make_pair(nullptr, 4);
+        P1 p1 = std::make_pair(nullptr, static_cast<short>(4));
         assert(p1.first == nullptr);
         assert(p1.second == 4);
     }
@@ -43,7 +43,7 @@ int main()
 #if TEST_STD_VER >= 14
     {
         typedef std::pair<int, short> P1;
-        constexpr P1 p1 = std::make_pair(3, 4);
+        constexpr P1 p1 = std::make_pair(3, static_cast<short>(4));
         static_assert(p1.first == 3, "");
         static_assert(p1.second == 4, "");
     }
